chore: created ReactorBase class and setup other implementations to use it
