+++ Testing/CarbonateTests/ReactorBaseTests.cs
@@ -1,0 +1,86 @@
+﻿// <copyright file="ReactorBaseTests.cs" company="KinsonDigital">
+// Copyright (c) KinsonDigital. All rights reserved.
+// </copyright>
+
+namespace CarbonateTests;
+
+using Carbonate.UniDirectional;
+using FluentAssertions;
+using Helpers.Fakes;
+using Xunit;
+
+public class ReactorBaseTests
+{
+    #region Method Tests
+    [Fact]
+    public void OnUnsubscribe_WhenNotUnsubscribed_InvokesAction()
+    {
+        // Arrange
+        var onReceiveInvoked = false;
+        void OnUnsubscribe() => onReceiveInvoked = true;
+
+        var sut = new ReactorBaseFake(Guid.NewGuid(), onUnsubscribe: OnUnsubscribe);
+
+        // Act
+        sut.OnUnsubscribe();
+
+        // Assert
+        onReceiveInvoked.Should().BeTrue();
+    }
+
+    [Fact]
+    public void OnUnsubscribe_WhenUnsubscribed_DoesNotInvokeActionAgain()
+    {
+        // Arrange
+        var totalInvokes = 0;
+        void OnUnsubscribe() => totalInvokes++;
+
+        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onUnsubscribe: OnUnsubscribe);
+        sut.OnUnsubscribe();
+
+        // Act
+        sut.OnUnsubscribe();
+
+        // Assert
+        totalInvokes.Should().Be(1);
+    }
+
+    [Fact]
+    public void OnError_WhenNotUnsubscribed_InvokesAction()
+    {
+        // Arrange
+        var onErrorInvoked = false;
+        void OnError(Exception ex) => onErrorInvoked = true;
+
+        var exception = new Exception("test-exception");
+
+        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onError: OnError);
+
+        // Act
+        sut.OnError(exception);
+
+        // Assert
+        onErrorInvoked.Should().BeTrue();
+    }
+
+    [Fact]
+    public void OnError_WhenNotSubscribed_DoesNotInvokeAction()
+    {
+        // Arrange
+        var onReceiveInvoked = false;
+        void OnError(Exception ex) => onReceiveInvoked = true;
+
+        var exception = new Exception("test-exception");
+
+        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onError: OnError);
+
+        sut.OnUnsubscribe();
+
+        // Act
+        sut.OnError(exception);
+
+        // Assert
+        onReceiveInvoked.Should().BeFalse();
+    }
+    #endregion
+}
