+++ Testing/CarbonateTests/UniDirectional/ReceiveReactorTests.cs
@@ -84,77 +84,6 @@ public class ReceiveReactorTests
             .WithMessage("The parameter must not be null. (Parameter 'message')");
     }
 
-    [Fact]
-    public void OnUnsubscribe_WhenNotUnsubscribed_InvokesAction()
-    {
-        // Arrange
-        var onReceiveInvoked = false;
-        void OnUnsubscribe() => onReceiveInvoked = true;
-
-        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onUnsubscribe: OnUnsubscribe);
-
-        // Act
-        sut.OnUnsubscribe();
-
-        // Assert
-        onReceiveInvoked.Should().BeTrue();
-    }
-
-    [Fact]
-    public void OnUnsubscribe_WhenUnsubscribed_DoesNotInvokeActionAgain()
-    {
-        // Arrange
-        var totalInvokes = 0;
-        void OnUnsubscribe() => totalInvokes++;
-
-        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onUnsubscribe: OnUnsubscribe);
-        sut.OnUnsubscribe();
-
-        // Act
-        sut.OnUnsubscribe();
-
-        // Assert
-        totalInvokes.Should().Be(1);
-    }
-
-    [Fact]
-    public void OnError_WhenNotUnsubscribed_InvokesAction()
-    {
-        // Arrange
-        var onErrorInvoked = false;
-        void OnError(Exception ex) => onErrorInvoked = true;
-
-        var exception = new Exception("test-exception");
-
-        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onError: OnError);
-
-        // Act
-        sut.OnError(exception);
-
-        // Assert
-        onErrorInvoked.Should().BeTrue();
-    }
-
-    [Fact]
-    public void OnError_WhenNotSubscribed_DoesNotInvokeAction()
-    {
-        // Arrange
-        var onReceiveInvoked = false;
-        void OnError(Exception ex) => onReceiveInvoked = true;
-
-        var exception = new Exception("test-exception");
-
-        var sut = new ReceiveReactor<int>(Guid.NewGuid(), onError: OnError);
-
-        sut.OnUnsubscribe();
-
-        // Act
-        sut.OnError(exception);
-
-        // Assert
-        onReceiveInvoked.Should().BeFalse();
-    }
-
     [Theory]
     [InlineData(null, "5739afd9-be4c-4402-a12d-6bcde35cc8c3")]
     [InlineData("", "5739afd9-be4c-4402-a12d-6bcde35cc8c3")]
